WIP - dpZipper from sqStickyLips
Start working in an update for stickyLips module from scratch.
We are adding a new module name dpZipper instead to edit the obsolete sqStickyLips.
Removing sqStickyLips from Extras.
For now we are able to select edges from polygon and generate upper, lower and middle curves.
We also can reverse curve direction based in the axis.