WIP #302 Limb auto rotate fix
#302 Leg bend autorotate issue.
Very hard to find a solution for that, but we got it after a long time.
We created a parent constraint from the main limb joint to the Ribbon locators group and also reviewed the constraints, removing middle orientConstraints and using parentConstraints skipping translations to have a good result without flippling.
Need to clean-up the code before publish it merging to master.