#287 Diamond name review
Naming review for #287 Dimond module. The correct name is Diamond. Thanks to Filipe Mello.
This update version is impacted by dictionary deletions. That means we could have problems with dpUpdateAutoInstall if we try to automatic update old versions.
If this occurs we need to delete all dpAutoRigSystem files manually and install the new version again, sorry.