WIP #061 set initial corner guide rotate
Added a new method to get rotate from corner zeroOut group, invert this value multiplying it to -1, and set the cornerLoc rotate with this result.
Need to develop a way to rotate the corner loc guide as a tangent of the elbow angle between shoulder and wrist.  Check if it works well for autoOrient method.