N545 Fixed Limb autoClavicle follow issue
Fixed Limb autoClavicle follow issue not orienting the scalable main group.
Instead of that we just used the shoulder null group to orient the articulation joint.
Shoulder null group was renamed to shoulder ref group, because it's now used to receive the Jax and Jar shoulder joints when we ask to have corrective joints.