N963-Controller Tag Add Parent
Added parent tag feature on Controller Tag checkout validator

ControllerTag Wiki:
Using the controllerTag is useful during the animation to make the rig faster, but it isn't good to apply during the rigging process, because it can cause some caching issues. So, we prefer to add them just when the rigging is done using this checkout validator.

Verify: It verifies if there are controllers without the controllerTag.
Fix: Add the controllerag to the controllers.
If it finds a json with controllers hierarchy from ControlsHierarchy, it will add the parent controller tag to the parent control.